Move schedulers directory to networks
