+++ tests/test_latent_diffusion_inferer.py
@@ -16,7 +16,7 @@ from parameterized import parameterized
 
 from generative.inferers import LatentDiffusionInferer
 from generative.networks.nets import VQVAE, AutoencoderKL, DiffusionModelUNet
-from generative.schedulers import DDPMScheduler
+from generative.networks.schedulers import DDPMScheduler
 
 TEST_CASES = [
     [
